Database tests fetch ligands
